% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{create_from_template}
\alias{create_from_template}
\title{Create a new file from an R Markdown template}
\usage{
create_from_template(template_name, output_file, output_dir = getwd())
}
\arguments{
\item{template_name}{Character string. The name of the template to use.
Should match the directory name in inst/rmarkdown/templates/.}

\item{output_file}{Character string. The name of the output file to create.}

\item{output_dir}{Character string. The directory where the file should be created.
Defaults to the current working directory.}
}
\value{
The path to the created file.
}
\description{
This function creates a new R Markdown file based on a specified template
in either the current directory or a specified path.
}
\examples{
\dontrun{
create_from_template("basic_stats_report", "my_analysis.Rmd")
create_from_template("advanced_analysis_report", "complex_analysis.Rmd", "path/to/project/")
}
}
